Split up 1400 lines test file.
